Call syncer in consensus (#513)
A node saves height from every ballot into map.
`map[/*/NodeAddress*/ string]/* height */uint64`
And when `BallotHeight` is greater than the node's latestConfirmedBlock.Height,
the node call `startSync`(it includes `syncer.SetSyncTargetBlock`).

The `BallotHeight` is the smallest of the heights above the threshold.
For example, if there are 5 nodes and threshold is 4 and height of 5 nodes are {10, 10, 3, 10, 10} then the `BallotHeight` is 10(https://github.com/bosnet/sebak/pull/513/files#diff-eb73fc2c4b3385428a4f522b335acbc1R9).

Node state transits from sync to consensus and consensus to sync each other
If the height from nodes exceeds the threshold, the node will be in the sync state.
If latestBlock is same as received ballot, then the node will be in the consensus state.

Sync with only ACCEPT and VotingYES ballot(#524)
And when the ballot height and block height in database are closed, call finishBallot in BallotCheckSYNC